Move PatternData into DisplayPatternLib
