+++ lib/DisplayPatternLib/src/DisplayPatternLib.h
@@ -4,5 +4,8 @@
 // This header file is only in place to link to the other files in the common lib.
 
 #include "PatternFactory.h"
+#include "PatternData.h"
+#include "enums\ColorPatternType.h"
+#include "enums\DisplayPatternType.h"
 
 #endif
