+++ lib/DisplayPatternLib/src/PatternData.cpp
