+++ lib/DisplayPatternLib/src/enums/ColorPatternType.h
