+++ lib/DisplayPatternLib/src/enums/DisplayPatternType.h
